% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precisionEst.R
\name{CovToPrecision}
\alias{CovToPrecision}
\title{CovToPrecision}
\usage{
CovToPrecision(Sigma, Q, Perm = NULL, calc.reo = TRUE)
}
\arguments{
\item{:}{Sigma       - (d x d) matrix the covariance matrix}

\item{:}{Q           - (d x d) symbolic matrix represnting graph structure}

\item{:}{Perm        - (d x 1) permutation list}

\item{:}{calc.reo    - (bool) if Perm is NULL and calc.reo is TRUE caclulate a reodering}
}
\value{
a sparse (d x d) matrix
}
\description{
verify Q, Sigma is symmetric, Perm is a vector of length D
Estimate a sparse Precision matrix (inverse covariance matrix)
using the Emperical covariance matrix and the sparsity structure of the
precision matrix.
}
\examples{
library(Matrix)
d <- 10
a <- 0.9
i <- c( 1:d, 1:(d-1), 2:d )
j <- c( 1:d, 2:d,     1:(d-1))
x <- c( c(1,rep(1 + a^2,d-2),1), rep(-a,d - 1), rep(-a, d - 1))
Q <- sparseMatrix( i = i, j = j, x = x)
Sigma <- as.matrix(solve(Q))
Qback <- CovToPrecision(Sigma, Q)

}

