% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precisionEst.R
\name{CovToPrecisionChol}
\alias{CovToPrecisionChol}
\title{CovToPrecisionChol}
\usage{
CovToPrecisionChol(Sigma, Q)
}
\arguments{
\item{:}{Sigma - (d x d) matrix the covariance matrix}

\item{:}{Q     - (d x d) symbolic matrix represnting graph structure}

\item{:}{Lcol  - (d x 1) Lcol[[i]] list of non-zero values for row i}
}

